ai-service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
psycopg2-binary==2.9.9
pandas==2.2.2
scikit-learn==1.5.0
numpy==1.26.4
python-dotenv==1.0.1
